--- frontend/src/components/Login/index.js
@@ -1,3 +1,0 @@
-import Login from './Login.cmp';
-
-export default Login;
