--- frontend/src/components/LoginStatus/index.js
@@ -1,3 +1,0 @@
-import StatusLogin from './StatusLogin';
-
-export default StatusLogin;
